A big mess of notebooks, but it'sokay
Had to export a bit of data from original notebook of random forest to reload it into new notebook "model_optimization.jpynb", where I fixed the code for the randomforest loop with integreated weather forecast. I did that so that I didn't have to run the whole notebook to work on the loop, which often lead to long waiting time everytime I had to restart the kernel.